Simplify API, move extensions from contract project
