+++ src/HonzaBotner.Discord.Services/Commands/PollCommands.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using DSharpPlus.CommandsNext;
 using DSharpPlus.CommandsNext.Attributes;
-using HonzaBotner.Discord.Extensions;
+using HonzaBotner.Discord.Services.Extensions;
 using HonzaBotner.Discord.Services.Commands.Polls;
 
 namespace HonzaBotner.Discord.Services.Commands
