+++ src/HonzaBotner.Discord.Services/Commands/ReminderCommands.cs
@@ -5,7 +5,7 @@ using Chronic.Core;
 using DSharpPlus.CommandsNext;
 using DSharpPlus.CommandsNext.Attributes;
 using DSharpPlus.Entities;
-using HonzaBotner.Discord.Extensions;
+using HonzaBotner.Discord.Services.Extensions;
 using HonzaBotner.Discord.Managers;
 using HonzaBotner.Discord.Services.Options;
 using HonzaBotner.Services.Contract;
